job_finder/requirements.txt
black==22.3.0
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Pillow==9.1.1
django-debug-toolbar==3.5.0
django-filter==22.1
celery==5.2.7

pydevd-pycharm==221.5921.27
